requirements.txt
Django==5.2.8
django-extensions==3.2.3
djangorestframework==3.15.2
sqlparse==0.5.1
djoser==2.3.3
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
django-allauth == 65.13.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0
whitenoise==6.7.0
requests==2.32.3
ruff==0.14.5
django-filter==25.2
django-axes==8.0.0
django_ratelimit==4.1.0
django_cleanup==9.0.0
daphne==4.2.1
channels==4.3.2
python-dotenv==1.1.0
google-genai==1.1.0